+++ ui-dist/assets/preferences-BpSeaUW7.js
@@ -1,4 +1,4 @@
-import{c}from"./createLucideIcon-DE2xCrS2.js";import{A as r,B as n}from"./api-BTFwMsQl.js";/**
+import{c}from"./createLucideIcon-2iIh_enc.js";import{A as r,B as n}from"./api-BlgFAraJ.js";/**
  * @license lucide-react v0.575.0 - ISC
  *
  * This source code is licensed under the ISC license.